simplifier: move if_expr related methods into separate file
These have tight coupling, and there is a sufficient number of them.
